### 2016-01-09 1f6e3e9
polishing for ropensci onboarding / cran submit
