+++ README.Rmd
@@ -9,13 +9,47 @@ output:
     toc: yes
 ---
 
-# Preface
-
 ![](logo/robotstxt.png)
 
 
+<table>
+<tr><td>&nbsp;</td></tr>
+<tr>
+<td> Ubuntu build </td>
+<td> <a href="https://travis-ci.org/petermeissner/robotstxt"><img src="https://api.travis-ci.org/petermeissner/robotstxt.svg?branch=master"></a></td>
+</tr>
+<tr><td>Version on CRAN  </td> 
+<td><a href="https://cran.r-project.org/web/packages/robotstxt/index.html"><img src="http://www.r-pkg.org/badges/version/robotstxt"></a></td></tr>
+<tr>
+<td>Downloads from <a href='http://cran.rstudio.com/'>CRAN.RStudio</a>&nbsp;&nbsp;&nbsp;</td>
+<td><img src="http://cranlogs.r-pkg.org/badges/grand-total/robotstxt"></td>
+</tr>
+<tr><td>&nbsp;</td></tr>
+</table>
 
-**Installation and Start**
+
+**Author:** Peter Meißner
+
+**Licence:** MIT
+
+**Description:**
+
+The package provides a class ('R6') and accompanying methods to
+parse and check 'robots.txt' files. Data fields are provided as 
+data frames and vectors. Permissions can be checked by providing
+path character vectors and optional bot names. 
+
+
+
+**Installation and start - stable version**
+
+```{r, eval=FALSE}
+install.packages("robotstxt")
+library(robotstxt)
+```
+
+
+**Installation and start - development version**
 
 ```{r, eval=FALSE}
 devtools::install_github("petermeissner/robotstxt")
@@ -23,111 +57,47 @@ library(robotstxt)
 ```
 
 
-**Documentation**
+**Robotstxt class documentation**
 
 ```{r, eval=FALSE}
 ?robotstxt
 ```
 
-
-
-
-# Example Usage 
-
-Loading the package and dplyr/magrittr for neater code  ... 
+**Usage**
 
 ```{r}
 library(robotstxt)
-library(dplyr)
+rtxt <- robotstxt$new(domain="wikipedia.org")
+rtxt$check("/")
 ```
 
-## The Object Oriented Style
+**Contribution**
 
-Initializing a new instance of the robotstxt class ... 
-```{r}
-rt_wikipedia <- robotstxt$new(domain="wikipedia.org")
+Please note that this project is released with a Contributor Code of Conduct. By participating in this project you agree to abide by its terms:
 
-rt_wikipedia
-```
-
-
-Having a look a some fields (the data stored inside the object) ... 
-
-... the domain for which the permissions apply ... 
-
-```{r}
-rt_wikipedia$domain
-```
-
-... the actual text of the robots.txt file ... 
-
-```{r}
-rt_wikipedia$text %>% 
-  substring(1, 400) 
-```
-
-... bots mentioned by name ...
-
-```{r}
-rt_wikipedia$bots
-```
-
-Checking if some/any bot might access the root path ... 
-
-```{r}
-# checking for access permissions
-rt_wikipedia$check(paths = c("/","api/"), bot = "*")
-rt_wikipedia$check(paths = c("/","api/"), bot = "Orthogaffe")
-rt_wikipedia$check(paths = c("/","api/"), bot = "Mediapartners-Google*  ")
-```
-
-## The Functional Style
-
-Retrieving robots.txt file ... 
-
-```{r}
-rtxt <- 
-  get_robotstxt("wikipedia.org") 
-
-rtxt %>% 
-  substring(1,400)
-``` 
-
-
[…]
