### 2016-01-10 #25 @sckott
Reviewers: @richfitz
