### 2016-01-10 #25 @Ironholds
1. Unless you own gmai.com there's a typo in your DESCRIPTION file ;p.
2. Do you need the alternate df implementations? They're fun but also a lot of extra code to replace stringsAsFactors = FALSE ;P.
3. The assignment [here](https://github.com/petermeissner/robotstxt/blob/master/R/permissions.R#L19) seems extraneous.
4. "did apply" to "applied" [here](https://github.com/petermeissner/robotstxt/blob/master/R/permissions.R#L123)

As you may notice, most of these quibbles are well, quibbles. On the whole the package is excellent :). I'll try to put time into a more detailed review later today or early tomorrow.
