+++ man/rt_get_rtxt.Rd
@@ -10,6 +10,7 @@ rt_get_rtxt(name = sample(rt_list_rtxt(), 1))
 \item{name}{name of the robots.txt files, defaults to a random drawn file ;-)}
 }
 \description{
-load robots.txt files saved along with the package
+load robots.txt files saved along with the package: 
+these functions ar very handy for testing (not used otherwise)
 }
 
