+++ man/rt_list_rtxt.Rd
@@ -7,6 +7,7 @@
 rt_list_rtxt()
 }
 \description{
-list robots.txt files saved along with the package
+list robots.txt files saved along with the package: 
+these functions ar very handy for testing (not used otherwise)
 }
 
