+++ DESCRIPTION
@@ -1,9 +1,9 @@
 Package: robotstxt
 Type: Package
-Title: A 'robots.txt' Parser and Webbot/Webspider/Webcrawler Permissions Checker
+Title: A 'robots.txt' Parser and 'Webbot'/'-spider'/'-crawler' Permissions Checker
 Version: 0.1.0
 Author: Peter Meissner
-Maintainer: Peter Meissner <retep.meissner@gmai.com>
+Maintainer: Peter Meissner <retep.meissner@gmail.com>
 Description: Class ('R6') and accompanying methods to
     parse and check 'robots.txt' files. Data fields are provided as
     data frames and vectors. Permissions can be checked by providing
