### 2016-01-11 #25 @petermeissner
Thanks for the ropensci infrastructure and thanks for the feedback - much-much-much appreciated.

# in response to @Ironholds

1) **Unless you own gmai.com there's a typo in your DESCRIPTION file ;p.**
- ... that is why devtools::build_win() never called back ... (Windows build only complains about qpdf missing [link1](http://win-builder.r-project.org/bB4FVpikrwIo/00check.log) [link2](http://win-builder.r-project.org/fH7n9bdiygk8/))
- **fixed**

2) **Do you need the alternate df implementations? They're fun but also a lot of extra code to replace stringsAsFactors = FALSE ;P.**
- First, I was inclined to say: "Yes, I can simply rewrite the code in the three places that are concerned - code simplicity over code fun."
- But then of course 5 tests broke telling me that functions were returning data.frames with factors inside.
- No problem I thought, that was expected and is easy to fix.
- Well it is easy to fix only for one instance ([link](https://github.com/petermeissner/robotstxt/blob/master/R/parse_robotstxt.R#L58)).
- The other instance ([link](https://github.com/petermeissner/robotstxt/blob/master/R/parse_robotstxt.R#L106), leading to 4 test failures) suffers from a bug described here: https://bugs.r-project.org/bugzilla/show_bug.cgi?id=16658 due to a problem with rbind.data.frame()
- I can solve it with something like this:
```r
for ( i in seq_len(dim(df)) ){
  if ( class(df[,i])=="factor" ){
    df <- as.character(df[,i])
  }
}
```
- **What do you think?**
- The [hellno](https://github.com/petermeissner/hellno) package and the `data.frame()`, `as.data.frame()`, `rbind()` implementations are part fun of cause but in the other hand they really solve a real problem that needs some constant thinking and hacking otherwise.
- despite my squabbling I am really open to advise

3) **The assignment here seems extraneous.**
- after checking extranous [here](http://dictionary.reference.com/browse/extraneous) and [here](http://www.thefreedictionary.com/extraneous) ...
- yes, developement relict
-  **fixed**

4) **"did apply" to "applied" here**
- **fixed**

# question of my own

`path_allowed()` ([link](https://github.com/petermeissner/robotstxt/blob/master/R/permissions.R#L118)) will response to unforseen permission constallations with a warning and return all data that led to the problem.

- Do you think that design decision is wise/ok ??
- Is it better to return a warning and `FALSE` for each path?
- Should it stop with an error? I think not - when used by headless programs this might break them also the problem is not fatal/severe
