+++ man/data.frame.Rd
@@ -1,21 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/hellno.R
-\name{data.frame}
-\alias{data.frame}
-\title{alternative data.frame() implementation}
-\usage{
-data.frame(..., stringsAsFactors = HELLNO)
-}
-\arguments{
-\item{...}{see \code{\link[base]{data.frame}}}
-
-\item{stringsAsFactors}{see \code{\link[base]{data.frame}} by default set to \code{FALSE}}
-}
-\description{
-\code{hellno::data.frame()} wraps up \code{base::data.frame()} so that
-\code{stringAsFactors} is set to \code{HELLNO} ( \code{== FALSE} ) by default
-}
-\seealso{
-\code{\link[base]{data.frame}}
-}
-
