+++ README.Rmd
@@ -28,6 +28,8 @@ output:
 </table>
 
 
+**Status:** feature complete, currently under code review by [ropensci](https://github.com/ropensci/onboarding/issues/25)
+
 **Author:** Peter Meißner
 
 **Licence:** MIT
