### 2016-01-12 #25 @Ironholds
On the whole: solid work! Some thoughts:

1. There are a lot of functions, such as <code>rt_get_fields</code>, which are documented but not exported. What this means in practice is that the code can't be found or used but the documentation does appear within the package index, which is pretty confusing. I'd suggest adding the "internal" roxygen2 keyword to the docs for those functions so that you don't index them.
2. The vignette is still using "Vignette Title" for indexing, so that's what CRAN will list it as (do not feel bad, I have made this mistake 150m times and never notice until I've already submitted it to CRAN and they've accepted >.>)
3. The vignette opens "The package provides a class (‘R6’)". This makes it sound like the class is called R6, at least to me.
4. The vignette could really benefit from a brief explanation of what robots.txt files are and why they matter at the top, with some examples, rather than just links at the bottom.
