### 2016-01-12 #25 @petermeissner
# in response to @Ironholds

2) **Do you need the alternate df implementations? They're fun but also a lot of extra code to replace stringsAsFactors = FALSE ;P.**

- Update: I did as suggested, **fixed**
