### 2016-01-12 31f0383
make internal functions internal
