+++ man/paths_allowed.Rd
@@ -21,4 +21,5 @@ wrapper to \code{\link{path_allowed}}
 \seealso{
 \link{path_allowed}
 }
+\keyword{internal}
 
