+++ man/sanitize_path.Rd
@@ -15,4 +15,5 @@ sanitized path
 \description{
 making paths uniform
 }
+\keyword{internal}
 
