+++ man/rt_get_fields_worker.Rd
@@ -19,4 +19,5 @@ fields}
 \description{
 extracting robotstxt fields
 }
+\keyword{internal}
 
