### 2016-01-12 1343bbb
ropensci onboarding vignette
