+++ vignettes/using_robotstxt.Rmd
@@ -1,21 +1,51 @@
 ---
-title: "using robotstxt"
+title: "Using Robotstxt"
 author: "Peter Meissner"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteIndexEntry{using_robotstxt}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
 
 # Description 
 
-The package provides a class ('R6') and accompanying methods to
+The package provides a robotstxt class ('R6') and accompanying methods to
 parse and check 'robots.txt' files. Data fields are provided as 
 data frames and vectors. Permissions can be checked by providing
 path character vectors and optional bot names. 
     
+    
+# Robots.txt files
+
+Robots.txt files are a way to kindly ask webbots, spiders, crawlers, wanderers and the like to access or not access certain parts of a webpage. The de facto  'standard' never made it beyond a informal ["Network Working Group INTERNET DRAFT"](http://www.robotstxt.org/norobots-rfc.txt). Nonetheless, the use of  robots.txt files is widespread (e.g. https://wikipedia.org/robots.txt, https://google.com/robots.txt) and bots from Google, Yahoo and the like will adhere to the rules defined in robots.txt files - although, their interpretation of those rules might differ. 
+
+As the name of the files already suggests robots.txt files are plain text and always found at the root of a domain. The syntax of the files in essence follows a `fieldname: value` scheme with optional preceding `user-agent: ...` lines to indicate the scope of the following rule block. Blocks are separated by blank lines and the omission of a user-agent field (which directly corresponds to the HTTP user-agent field) is seen as referring to all bots. `#` serves to comment lines and parts of lines. Everything after `#` until the end of line is regarded a comment. Possible field names are: user-agent, disallow, allow, crawl-delay, sitemap, and host. 
+
+
+Let us have an example file to get an idea how a robots.txt file might look like. The file below starts with a comment line followed by a line disallowing access to any content -- everything that is contained in root ("`/`") -- for all bots. The next block concerns GoodBot and NiceBot. Those two get the previous permissions lifted by being disallowed nothing. The third block is for PrettyBot. PrettyBot likes shiny stuff and therefor gets a special permission for everything contained in the "`/shinystuff/`" folder while all other restrictions still hold. In the last block all bots are asked to pause at least 5 seconds between two visits. 
+
+
+```robots.txt
+# this is a comment 
+# a made up example of an robots.txt file
+
+Disallow: /
+
+User-agent: GoodBot # another comment
+User-agent: NiceBot
+Disallow: 
+
+User-agent: PrettyBot
+Allow: /shinystuff/
+
+Crawl-Delay: 5
+```
+
+For more information have a look at: http://www.robotstxt.org/norobots-rfc.txt, where the robots.txt file 'standard' is described formally. Valuable introductions can be found at http://www.robotstxt.org/robotstxt.html as well as at https://en.wikipedia.org/wiki/Robots_exclusion_standard - of cause. 
+
+
 
 # Example Usage 
 
@@ -31,7 +61,11 @@ library(dplyr)
[…]
