+++ DESCRIPTION
@@ -4,7 +4,7 @@ Title: A 'robots.txt' Parser and 'Webbot'/'-spider'/'-crawler' Permissions Check
 Version: 0.1.0
 Author: Peter Meissner
 Maintainer: Peter Meissner <retep.meissner@gmail.com>
-Description: Class ('R6') and accompanying methods to
+Description: 'Robotstxt' ('R6') class and accompanying methods to
     parse and check 'robots.txt' files. Data fields are provided as
     data frames and vectors. Permissions can be checked by providing
     path character vectors and optional bot names.
