### 2016-01-12 #25 @petermeissner
# in response to @Ironholds II

1) **drop unexported from manual**
  - **done**

2) **vignette title / index entry**
- **done**

3) **The vignette opens "The package  ...**
- **done**

4) **The vignette could really benefit from a brief explanation of what robots.txt files are and why they matter ...**
- **done**
