### 2016-01-13 07a6f39
ropensci onboarding: more informative permission error warning and return of NA when problems occur
