### 2016-01-13 c05ab78
more informative warning if permission deduction fails
