### 2016-01-13 #25 @petermeissner
# in regard to how path_allowed() handles problems with permission deduction

- path_allowed() will now warn more informative describing what happened:
> robotstxt::path_allowed() says:
> Encountered problems while trying to determine bot permissions, returning NA instead of TRUE/FALSE

- furthermore, path_allowed() will return NA in cases where the permission cannot be determined
