+++ DESCRIPTION
@@ -1,9 +1,10 @@
 Package: robotstxt
 Type: Package
 Title: A 'robots.txt' Parser and 'Webbot'/'-spider'/'-crawler' Permissions Checker
-Version: 0.1.0
-Author: Peter Meissner
-Maintainer: Peter Meissner <retep.meissner@gmail.com>
+Version: 0.1.1
+Authors@R: c(person("Peter", "Meissner", role = c("aut", "cre"),
+		     email = "retep.meissner@gmail.com"),
+	      person("Oliver", "Keys", role = "ctb"))
 Description: 'Robotstxt' ('R6') class and accompanying methods to
     parse and check 'robots.txt' files. Data fields are provided as
     data frames and vectors. Permissions can be checked by providing
