### 2016-02-07 6a96caf
CRAN submission
