+++ README.md
@@ -12,7 +12,7 @@ Peter Meißner
 <td> <a href="https://travis-ci.org/petermeissner/robotstxt"><img src="https://api.travis-ci.org/petermeissner/robotstxt.svg?branch=master"></a></td>
 </tr>
 <tr><td>Version on CRAN  </td> 
-<td><a href="https://cran.r-project.org/web/packages/robotstxt/index.html"><img src="http://www.r-pkg.org/badges/version/robotstxt"></a></td></tr>
+<td><a href="http://cran.r-project.org/package=robotstxt"><img src="http://www.r-pkg.org/badges/version/robotstxt"></a></td></tr>
 <tr>
 <td>Downloads from <a href='http://cran.rstudio.com/'>CRAN.RStudio</a>&nbsp;&nbsp;&nbsp;</td>
 <td><img src="http://cranlogs.r-pkg.org/badges/grand-total/robotstxt"></td>
@@ -21,9 +21,11 @@ Peter Meißner
 </table>
 
 
-**Status:** feature complete, currently under code review by [ropensci](https://github.com/ropensci/onboarding/issues/25)
+
 
 **Author:** Peter Meißner
+
+**Contributer:** Oliver Keys (code review)
 
 **Licence:** MIT
 
@@ -77,31 +79,7 @@ rtxt$check("/")
 
 **Contribution**
 
-Please note that this project is released with a Contributor Code of Conduct. By participating in this project you agree to abide by its terms:
-
-> As contributors and maintainers of this project, we pledge to respect all people who 
-contribute through reporting issues, posting feature requests, updating documentation,
-submitting pull requests or patches, and other activities.
-> 
-> We are committed to making participation in this project a harassment-free experience for
-everyone, regardless of level of experience, gender, gender identity and expression,
-sexual orientation, disability, personal appearance, body size, race, ethnicity, age, or religion.
-> 
-> Examples of unacceptable behavior by participants include the use of sexual language or
-imagery, derogatory comments or personal attacks, trolling, public or private harassment,
-insults, or other unprofessional conduct.
-> 
-> Project maintainers have the right and responsibility to remove, edit, or reject comments,
-commits, code, wiki edits, issues, and other contributions that are not aligned to this 
-Code of Conduct. Project maintainers who do not follow the Code of Conduct may be removed 
-from the project team.
-> 
-> Instances of abusive, harassing, or otherwise unacceptable behavior may be reported by 
-opening an issue or contacting one or more of the project maintainers.
-> 
-> This Code of Conduct is adapted from the Contributor Covenant 
-(http:contributor-covenant.org), version 1.0.0, available at 
-http://contributor-covenant.org/version/1/0/0/
+Please note that this project is released with a Contributor Code of Conduct. By participating in this project you agree to abide by its terms: http://contributor-covenant.org/version/1/0/0/
 
 
 
