+++ README.Rmd
@@ -19,7 +19,7 @@ output:
 <td> <a href="https://travis-ci.org/petermeissner/robotstxt"><img src="https://api.travis-ci.org/petermeissner/robotstxt.svg?branch=master"></a></td>
 </tr>
 <tr><td>Version on CRAN  </td> 
-<td><a href="https://cran.r-project.org/web/packages/robotstxt/index.html"><img src="http://www.r-pkg.org/badges/version/robotstxt"></a></td></tr>
+<td><a href="http://cran.r-project.org/package=robotstxt"><img src="http://www.r-pkg.org/badges/version/robotstxt"></a></td></tr>
 <tr>
 <td>Downloads from <a href='http://cran.rstudio.com/'>CRAN.RStudio</a>&nbsp;&nbsp;&nbsp;</td>
 <td><img src="http://cranlogs.r-pkg.org/badges/grand-total/robotstxt"></td>
@@ -31,6 +31,8 @@ output:
 **Status:** feature complete, currently under code review by [ropensci](https://github.com/ropensci/onboarding/issues/25)
 
 **Author:** Peter Meißner
+
+**Contributer:** Oliver Keys (code review)
 
 **Licence:** MIT
 
