+++ DESCRIPTION
@@ -1,11 +1,11 @@
 Package: robotstxt
 Type: Package
-Title: A 'robots.txt' Parser and 'Webbot'/'-spider'/'-crawler' Permissions Checker
-Version: 0.1.1
+Title: A 'robots.txt' Parser and 'Webbot'/'Spider'/'Crawler' Permissions Checker
+Version: 0.1.2
 Authors@R: c(person("Peter", "Meissner", role = c("aut", "cre"),
 		     email = "retep.meissner@gmail.com"),
 	      person("Oliver", "Keys", role = "ctb"))
-Description: 'Robotstxt' ('R6') class and accompanying methods to
+Description: Provides a 'Robotstxt' ('R6') class and accompanying methods to
     parse and check 'robots.txt' files. Data fields are provided as
     data frames and vectors. Permissions can be checked by providing
     path character vectors and optional bot names.
