### 2016-02-08 516a2fa
CRAN release
