+++ README.Rmd
@@ -13,7 +13,6 @@ output:
 
 
 <table>
-<tr><td>&nbsp;</td></tr>
 <tr>
 <td> Ubuntu build </td>
 <td> <a href="https://travis-ci.org/petermeissner/robotstxt"><img src="https://api.travis-ci.org/petermeissner/robotstxt.svg?branch=master"></a></td>
@@ -24,7 +23,6 @@ output:
 <td>Downloads from <a href='http://cran.rstudio.com/'>CRAN.RStudio</a>&nbsp;&nbsp;&nbsp;</td>
 <td><img src="http://cranlogs.r-pkg.org/badges/grand-total/robotstxt"></td>
 </tr>
-<tr><td>&nbsp;</td></tr>
 </table>
 
 
