### 2016-02-11 #25 @richfitz
I'm sorry, I have left this very late so I will just jot down the thoughts I have rather than a full review, in the hope that it is useful.

My biggest major comment is that I don't see what R6 is adding here; it seems that the return value could be a list (which contains one function).

Don't get me wrong -- I *love* R6.  But in cases where reference semantics aren't needed or being used it seems an unnecessary weirdness for most users (and in my experience users do find them weird).  See the code below for a pure-list implementation of your function (purposefully left very similar) to show that there is no real difference in implementation.  The only difference is that initialisation looks less weird (`x <- robotstxt()` not `x <- robotstxt$new()`).

```r
robotstxt <- function(domain, text) {
  ## check input
  self <- list()
  if (missing(domain)) {
    self$domain <- NA
  }
  if (!missing(text)){
    self$text <- text
    if(!missing(domain)){
      self$domain <- domain
    }
  }else{
    if(!missing(domain)){
      self$domain <- domain
      self$text   <- get_robotstxt(domain)
    }else{
      stop("robotstxt: I need text to initialize.")
    }
  }
  ## fill fields with default data

  tmp <- parse_robotstxt(self$text)
  self$bots        <- tmp$useragents
  self$comments    <- tmp$comments
  self$permissions <- tmp$permissions
  self$crawl_delay <- tmp$crawl_delay
  self$host        <- tmp$host
  self$sitemap     <- tmp$sitemap
  self$other       <- tmp$other

  self$check <- function(paths="/", bot="*", permission=self$permissions) {
    paths_allowed(permissions=permission, paths=paths, bot=bot)
  }

  class(self) <- "robotstxt"
  self
}
```

(this passes checks in the package by deleting only the `$new` in the test file).

I really like the idea of downloading the file once and testing repeatedly.  I wonder if that could be extended with the assumption that the file does not change during an R session.  Then one could memoise the calls and avoid having to have any sort of object for the users to worry about.

```r
cache <- new.env(parent=emptyenv())
path_allowed <- function(url) {
  x <- httr::parse_url(url)
  obj <- cache[[x$hostname]]
  if (is.null(obj)) {
    cache[[x$hostname]] <- obj <- robotstxt(hostname)
  }
  obj$check(x$path)
}
```

# Minor comments

Avoid `httr::content(rtxt)` in package code (see `?httr::content`) in favour of checking that the return type was correct.  In development versions of httr I find I also have to declare `encoding="UTF-8"` to avoid a message on every request, too.

In `get_robotstxt`, is `warning` the appropriate behaviour?  I would have thought that for 403 and 5xx errors it should be an error.  For 404 errors could we not assume that there *is_* no robots.txt and therefore it's all fair game?

I find that `missing()` lends itself to hard-to-diagnose errors when functions are several layers deep; consider explicit NULL arguments and `is.null()` tests for missing-ness (purely style).

You have a partial match of `$useragent` to `$useragents` in `R/permissions.R` and throughout `tests/testthat/test_parser.R`
