### 2016-03-20 petermeissner/robotstxt#5 @petermeissner
get_robotstxt redponse to HTTP errors
https://github.com/ropensci/onboarding/issues/25 by @richfitz

> In get_robotstxt, is warning the appropriate behaviour? I would have thought that for 403 and 5xx errors it should be an error. For 404 errors could we not assume that there is_ no robots.txt and therefore it's all fair game?
