### 2016-03-20 petermeissner/robotstxt#9 @petermeissner
do not use missing for parameter check
ropensci/onboarding#25 by @richfitz

> I find that missing() lends itself to hard-to-diagnose errors when functions are several layers deep; consider explicit NULL arguments and is.null() tests for missing-ness (purely style).
