### 2016-03-20 petermeissner/robotstxt#10 @petermeissner
partial match useragent / useragents
ropensci/onboarding#25 by @richfitz

> You have a partial match of $useragent to $useragents in R/permissions.R and throughout tests/testthat/test_parser.R
