### 2016-03-20 petermeissner/robotstxt#11 @petermeissner
declare encoding="UTF-8"
ropensci/onboarding#25 by @richfitz

> Avoid httr::content(rtxt) in package code (see ?httr::content) in favour of checking that the return type was correct. In development versions of httr I find I also have to declare encoding="UTF-8" to avoid a message on every request, too.
