+++ README.Rmd
@@ -30,7 +30,7 @@ output:
 
 **Author:** Peter Meißner
 
-**Contributer:** Oliver Keys (code review)
+**Contributer:** Oliver Keys (code review), Rich FitzJohn (code review)
 
 **Licence:** MIT
 
