### 2016-03-20 petermeissner/robotstxt#5 @petermeissner
- 404 -> yes that should be the interpretation
