### 2016-03-21 petermeissner/robotstxt#10 @petermeissner
thats right and no good
