### 2016-03-21 petermeissner/robotstxt#11 @petermeissner
1 out of 10 problems is related to encodings
