+++ tests/testthat/test_parser.R
@@ -24,19 +24,19 @@ rtxt_host  <- rt_get_rtxt("host.txt")
 
 test_that(
   "all user agents are extracted", {
-    expect_true(all( parse_robotstxt(rtxt_asb   )$useragent %in% c("*", "Google") ))
-    expect_true(all( parse_robotstxt(rtxt_dafa  )$useragent %in% c("*") ))
-    expect_true(all( parse_robotstxt(rtxt_dafbb )$useragent %in% c("BadBot") ))
-    expect_true(all( parse_robotstxt(rtxt_dsfa  )$useragent %in% c("*") ))
-    expect_true(all( length(parse_robotstxt(rtxt_empty )$useragent) == 0  ))
-    expect_true(all( parse_robotstxt(rtxt_amzn  )$useragent %in% c("EtaoSpider", "*") ))
-    expect_true(all( parse_robotstxt(rtxt_bt    )$useragent %in% c("*") ))
-    expect_true(all( parse_robotstxt(rtxt_ggl   )$useragent %in% c("*") ))
-    expect_true(all( parse_robotstxt(rtxt_nyt   )$useragent %in% c("*", "Mediapartners-Google", "AdsBot-Google", "adidxbot" ) ))
-    expect_true(all( parse_robotstxt(rtxt_spgl  )$useragent %in% c("WebReaper", "Slurp") ))
-    expect_true(all( parse_robotstxt(rtxt_yh    )$useragent %in% c("*") ))
-    expect_true(all( parse_robotstxt(rtxt_she   )$useragent %in% c("*","mein-Robot", "UniversalRobot/1.0") ))
-    expect_true(all( parse_robotstxt(rtxt_datao )$useragent %in% c("BadBot","Googlebot") ))
+    expect_true(all( parse_robotstxt(rtxt_asb   )$useragents %in% c("*", "Google") ))
+    expect_true(all( parse_robotstxt(rtxt_dafa  )$useragents %in% c("*") ))
+    expect_true(all( parse_robotstxt(rtxt_dafbb )$useragents %in% c("BadBot") ))
+    expect_true(all( parse_robotstxt(rtxt_dsfa  )$useragents %in% c("*") ))
+    expect_true(all( length(parse_robotstxt(rtxt_empty )$useragents) == 0  ))
+    expect_true(all( parse_robotstxt(rtxt_amzn  )$useragents %in% c("EtaoSpider", "*") ))
+    expect_true(all( parse_robotstxt(rtxt_bt    )$useragents %in% c("*") ))
+    expect_true(all( parse_robotstxt(rtxt_ggl   )$useragents %in% c("*") ))
+    expect_true(all( parse_robotstxt(rtxt_nyt   )$useragents %in% c("*", "Mediapartners-Google", "AdsBot-Google", "adidxbot" ) ))
+    expect_true(all( parse_robotstxt(rtxt_spgl  )$useragents %in% c("WebReaper", "Slurp") ))
+    expect_true(all( parse_robotstxt(rtxt_yh    )$useragents %in% c("*") ))
+    expect_true(all( parse_robotstxt(rtxt_she   )$useragents %in% c("*","mein-Robot", "UniversalRobot/1.0") ))
+    expect_true(all( parse_robotstxt(rtxt_datao )$useragents %in% c("BadBot","Googlebot") ))
   }
 )
 
