### 2016-03-21 petermeissner/robotstxt#9 @petermeissner
... hmmm, to prevent the function using an object from outside the function - right. good idea.
