### 2016-03-22 petermeissner/robotstxt#7 @petermeissner
Caching of robots.txt file downloads is a very good idea - those things should not change within a sessions time.

But, I did not put the caching in path allowed but in `get_robotstxt()` because that is the place where the download is supposed to happen.
