+++ README.md
@@ -71,11 +71,11 @@ rtxt$check("/")
 ```
 
 ```
-##    / 
-## TRUE
+##     / 
+## FALSE
 ```
 
-**Contribution**
+**Contribution - aka the think-twice-be-nice-rule**
 
 Please note that this project is released with a Contributor Code of Conduct. By participating in this project you agree to abide by its terms:
 
