+++ README.Rmd
@@ -73,7 +73,7 @@ rtxt <- robotstxt$new(domain="wikipedia.org")
 rtxt$check("/")
 ```
 
-**Contribution**
+**Contribution - aka the think-twice-be-nice-rule**
 
 Please note that this project is released with a Contributor Code of Conduct. By participating in this project you agree to abide by its terms:
 
