### 2016-03-23 74657d3
fixes #6 #7
