+++ R/robotstxt.R
@@ -1,126 +1,87 @@
 
 
-#' An object representation of robots.txt files
-#'
-#' @docType class
-#'
-#' @name robotstxt
+#' Generate object representations of a robots.txt file
 #'
 #' @export
-#' @importFrom R6 R6Class
 #'
 #' @keywords data
 #'
-#' @return Object of \code{\link{R6Class}} with method(s) for bot permission checking.
+#' @return Object (list) of class robotstxt with parsed data from a
+#'   robots.txt file and method(s) for bot permission checking.
 #'
-#' @format \code{\link{R6Class}} object.
+#' @field domain character vector holding domain name for which the robots.txt
+#'   file is valid; will be set to NA if not supplied on initialization
 #'
-#' @field domain
-#'  character vector holding domain name for which the robots.txt file is valid;
-#'  will be set to NA if not supplied on initialization
+#' @field text character vector of text of robots.txt file; either supplied on
+#'   initializetion or automatically downloaded from domain supplied on
+#'   initialization
 #'
-#' @field text
-#'  character vector of text of robots.txt file;
-#'  either supplied on initializetion
-#'  or automatically downloaded from domain supplied on initialization
+#' @field bots character vector of bot names mentionend in robots.txt
 #'
-#' @field bots
-#'  character vector of bot names mentionend in robots.txt
+#' @field permissions data.frame of bot permissions found in robots.txt file
 #'
-#' @field permissions
-#'  data.frame of bot permissions found in robots.txt file
+#' @field host data.frame of host fields found in robots.txt file
 #'
-#' @field host
-#'  data.frame of host fields found in robots.txt file
+#' @field sitemap data.frame of sitemap fields found in robots.txt file
 #'
-#' @field sitemap
-#'  data.frame of sitemap fields found in robots.txt file
+#' @field other data.frame of other - none of the above - fields found in
+#'   robots.txt file
 #'
-#' @field other
-#'  data.frame of other - none of the above - fields found in robots.txt file
+#' @usage robotstxt(domain="mydomain.com")
+#' robotstxt(text="User-agent: *\nDisallow: /")
 #'
-#'
-#' @section Methods:
-#' \describe{
-#'  \item{\code{
-#'    initialize(domain, text) }}{
-#'    Method called when initialising object via
-#'    \code{robotstxt$new()}. Needs either \code{text} or \code{domain} to be
-#'    present at initialization. If only domain is supplied -- should be seen as
-#'    default -- than the robots.txt file for that domain will be downloaded. If
-#'    \code{text} is supplied as well, nothing will be downloaded. If only
-#'    \code{text} is supplied than domain is set to '???'.
-#'  }
-#'  \item{\code{
-#'    check( path="/", bot="*" ) }}{
-#'    Method for checking whether or not paths are allowed to be accessed by a bot
-#'  }
-#' }
 #'
 #' @examples
 #' \dontrun{
-#' rt <- robotstxt$new(domain="google.com")
+#' rt <- robotstxt(domain="google.com")
 #' rt$bots
 #' rt$permissions
 #' rt$check( paths = c("/", "forbidden"), bot="*")
 #' }
 #'
-robotstxt <-
-  R6::R6Class(
-  # class name
-    "robotstxt",
-  # declaration
-    public = list(
-  # puplic data fields
-      domain      =  NA,
-      text        =  NA,
-      bots        =  NA,
-      comments    =  NA,
-      permissions =  NA,
[…]
