+++ R/permissions.R
@@ -59,7 +59,6 @@ sanitize_permissions <- function(permissions){
 #' @keywords internal
 
 path_allowed <- function(permissions, path="/", bot="*"){
-
   # checking and initializetion
   stopifnot(length(bot)==1)
   if( is.null(bot) | bot=="" | is.na(bot) ) bot <- "*"
@@ -134,18 +133,75 @@ path_allowed <- function(permissions, path="/", bot="*"){
   return(NA)
 }
 
-#' check if a bot has permissions to access page
+
+#' check if a bot has permissions to access page(s)
 #'
 #' wrapper to \code{\link{path_allowed}}
-#' @param permissions data.frame generated by \code{rt_get_permissions()} with
-#'  three columns: 'useragent' (name of the bot), 'permission' (permission
-#'  granted: Disallow/Allow) and 'value' (path for which permission applies)
+#' @param domain Domain for which paths should be checked. Defaults to "auto".
+#'   If set to "auto" function will try to guess the domain by parsing the paths
+#'   argument. Note however, that these are educated guesses which might utterly
+#'   fail. To be on the save side, provide appropriate domains manually.
 #' @param bot name of the bot, defaults to "*"
 #' @param paths paths for which to check bot's permission, defaults to "/"
 #' @seealso \link{path_allowed}
 #' @export
-#' @keywords internal
-paths_allowed <- function(permissions, paths="/", bot="*"){
-  sapply(paths, path_allowed, permissions=permissions, bot=bot)
+paths_allowed <- function(paths="/", domain="auto", bot="*"){
+  # dev # paths <- c("https://pmeissner.com/_layouts/index.html","https://wikipedia.org", "http://pmeissner.com/index.html")
+  if( all(domain == "auto") ){
+    domain <- guess_domain(paths)
+    paths  <- remove_domain(paths)
+  }
+  if( length(unique(domain))==1 ){
+    domain <- domain[1]
+  }
+  permissions <-
+    lapply(domain, function(x){robotstxt(x)$permissions} )
+  worker <- function(path, permissions, bot, domain){
+    if( is.na(domain) ){
+      return(NA)
+    }else{
+      path_allowed(permissions = permissions, path = path, bot = bot)
+    }
+  }
+  tmp <- mapply(worker, path=paths, permissions=permissions, bot=bot, domain)
+  names(tmp) <- NULL
+  return(tmp)
 }
+
+
+#' function guessing domain from path
+#' @param x path aka URL from which to infer domain
+guess_domain <- function(x){
+  if(length(x)>1){
+    return(unlist(lapply(x, guess_domain)))
+  }else{
+    hostname <- httr::parse_url(x)$hostname
+    if( is.null(hostname) ){
+      hostname <- NA
+    }
+    return(hostname)
+  }
+}
+
+#' function to remove domain from path
+#' @param x path aka URL from which to first infer domain and then remove it
+remove_domain <- function(x){
+  unlist(lapply(
+      x,
+      function(x){
+        if( is.na(x) ){
+          return(x)
+        }else{
+          stringr::str_replace(x, paste0("^.*", guess_domain(x)), "")
+        }
+      }
+  ))
+}
+
+
+
+
+
+
+
 
