+++ man/remove_domain.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/permissions.R
+\name{remove_domain}
+\alias{remove_domain}
+\title{function to remove domain from path}
+\usage{
+remove_domain(x)
+}
+\arguments{
+\item{x}{path aka URL from which to first infer domain and then remove it}
+}
+\description{
+function to remove domain from path
+}
+
