+++ man/print.robotstxt.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/robotstxt.R
+\name{print.robotstxt}
+\alias{print.robotstxt}
+\title{printing robotstxt}
+\usage{
+\method{print}{robotstxt}(x, ...)
+}
+\arguments{
+\item{x}{robotstxt instance to be printed}
+
+\item{...}{goes down the sink}
+}
+\description{
+printing robotstxt
+}
+
