### 2016-03-31 #25 @petermeissner
Just some further thoughts ...

In regard to discarding R6 because of not using "reference semantics" I have done some research - [ah, that's what it is](http://cis-linux1.temple.edu/~yates/cis1068/sp12/slides/objects-as-parameters.pdf).

- First, I will still leave R6 out because it can be done without - as you have suggested - and its one dependency less.
- But, using reference semantics is not the only reason to use Oo-design.
- In this case I think the tight connection of data and functions would be reason enough to use it.
- In regard to users who might be confused by using `robotstxt$new()` instead of `robotstxt()`: I do not think that the first is more confusing than the latter. In the first case the user is explicitly told that he will use something that might be unusual. In the second case it is more like giving the user that looks like something he understands but than letting it do black magic.
