+++ DESCRIPTION
@@ -2,8 +2,11 @@ Package: robotstxt
 Type: Package
 Title: A 'robots.txt' Parser and 'Webbot'/'Spider'/'Crawler' Permissions Checker
 Version: 0.3.1.90000
-Authors@R: c(person("Peter", "Meissner", role = c("aut", "cre"),
-		     email = "retep.meissner@gmail.com"),
+Authors@R: c(
+        person(
+          "Peter", "Meissner", role = c("aut", "cre"),
+          email = "retep.meissner@gmail.com"
+        ),
 	      person("Oliver", "Keys", role = "ctb"),
 	      person("Rich", "FitzJohn", role = "ctb")
 	      )
