### 2016-04-27 #25 @sckott
Great! Thanks. Approved.

I added you to a team with admin access, and now you should be able to transfer the repo to `ropenscilabs` - we move pkgs to `ropensci` github org once more mature - your pkg is still a fully ropensci pkg even though in `ropenscilabs`
