### 2016-04-27 #25 @sckott
> It is done via Settings > Danger Zone > Transfer ownership : "ropenscilabs"

yes

> How is pushing to CRAN handled thereafter? I would like to push the current thing to CRAN.

same as before - you are still the maintainer of your pkg, and can submit to CRAN as you like. You can push new version to CRAN before transferring if you like
