### 2016-04-27 4f5f249
ropensci accepted
