+++ cran-comments.md
@@ -1,41 +1,16 @@
 ## Test environments
 
 * Ubuntu 12.04 (on travis-ci: old, current, devel; https://travis-ci.org/petermeissner/robotstxt)
-* LinuxMint 17.3 (local, current)
+* Ubuntu 14.04 LTS (local, current)
 * win-builder (devel and release)
 
 
 ## R CMD check results
 
-0 errors | 0 warnings | 1 note
-
-* This is a new release.
+0 errors | 0 warnings | 0 notes
 
 
-## Note
 
-* checking CRAN incoming feasibility ... NOTE
-Maintainer: ‘Peter Meissner <retep.meissner@gmail.com>’
-New submission
-
-License components with restrictions and base license permitting such:
-  MIT + file LICENSE
-File 'LICENSE':
-  YEAR: 2016
-  COPYRIGHT HOLDER: Peter Meissner
-Found the following (possibly) invalid URLs:
-  URL: http://cran.r-project.org/package=robotstxt
-    From: README.md
-    Status: 404
-    Message: Not Found
-
---> Te license is MIT, one of the valid ones for CRAN
---> The link is in canonical form and should be valid when the package is published on CRAN
-
-
-## Reverse dependencies
-
-This is a new release, so there are no reverse dependencies.
 
 
 
