+++ man/robotstxt.Rd
@@ -49,9 +49,6 @@ domains root and no bot in particular. check() has two arguments:
 paths and bot. The first is for supplying the paths for which to check
 permissions and the latter to put in the name of the bot.}
 }}
-\section{return values}{
-
-}
 \examples{
 \dontrun{
 rt <- robotstxt(domain="google.com")
