### 2016-04-27 6290bf6
preparing CRAN submission
