+++ cran-comments.md
@@ -7,10 +7,12 @@
 
 ## R CMD check results
 
-0 errors | 0 warnings | 0 notes
+0 errors | 0 warnings | 1 notes
 
 
 
+## Comments
 
+There are links in the DESCRIPTION file that will not work at the moment of submission. Those will work however work asap (max 1 day) after the submission is accepted since  GitHub repository will move to to RopenSciLabs and than links shall be fine.
 
 
