### 2016-04-27 2c8295b
preparing CRAN submission
