+++ R/robotstxt.R
@@ -116,19 +116,19 @@ print.robotstxt <- function(x, ...){
   }
   if( length(tmp$bots) > 7 ){
     tmp$bots <-
-      c(head(tmp$bots), "", paste("[... ", length(tmp$bots)-5, "items omitted ...]") )
+      c(utils::head(tmp$bots), "", paste("[... ", length(tmp$bots)-5, "items omitted ...]") )
   }
   for(i in c("permissions", "crawl_delay", "host", "sitemap", "other") ){
     if( dim(tmp[[i]])[1] > 7  ){
       tmp_fill <- data.frame(cbind( paste("[... ", dim(tmp[[i]])[1]-5, "items omitted ...]"), "",""))
       names(tmp_fill) <- names(tmp[[i]])
-      tmp[[i]]  <- rbind( head(tmp[[i]]), "", tmp_fill )
+      tmp[[i]]  <- rbind( utils::head(tmp[[i]]), "", tmp_fill )
     }
   }
   if( dim(tmp$comments)[1] > 7 ){
     tmp_fill <- data.frame(cbind( "", paste("[... ", dim(tmp[["comments"]])[1]-5, "items omitted ...]")))
     names(tmp_fill) <- names(tmp[["comments"]])
-    tmp[["comments"]]  <- rbind( head(tmp[["comments"]]), "", tmp_fill )
+    tmp[["comments"]]  <- rbind( utils::head(tmp[["comments"]]), "", tmp_fill )
   }
   print.default(tmp)
   invisible(x)
